update environment check if move is valid and add state reset to init
